Settings setup and Logging
This patch adds a configuration file (config.yaml) and a parser for it.
It also parses the command line arguments, and overwrites the system settings
from config.yaml in the case that command line duplicates any options.

Python's logging is being used to log the expected exceptions.

PyYAML was added as dependency in the README.rst