Introduces BuildManager and classes for distribution abstraction
BuildManager will be responsible to deal with all the workflow necessary from
the distro detection till the build of the packages.

This patch also include utilities to detect the GNU/linux distribution and
abstractions for CentOS, the first one we support.

Finally, our own set of exceptions was included. The idea here is to set a
proper fault message and return code. At this momment we return 1 for whatever
kind of error that happens and 0 for success.