Introduces Packages abstraction
Package class is the object that will gather all the data regarding the packages
to build for the detected distribution, like clone url of the project's git tree
and their spec for the supported GNU/Linux distribution.

This also includes new exceptions for dealing w/ package errors and a
refactoring on the exception inheritance.

An extra directory was included to specify centOS's version and the yaml
descriptors got the spec data added.