Add mock to build on centOS
This commit refactors a lot of things, so step by step:
    - As the commit title says: Mock class was added it is responsible for
calling mock though subprocess.Popen. Note: it requires further refactoring.
    - Added mock configuration file for ppc64el at
extras/centOS/7.2/mock/epel-7-ppc64le.cfg. That was also contributed to mock
upstream, released, but not yet present on the distro, thus I've added it here
and customized it to not check for the GPG keys. Apparently the GPG keys need to
be on the system instead of the chroot. It's on the TODO list to see if we can
workaround this.
    - Auxiliar files necessary to create the RPM were moved from, for instance
for the kernel, components/kernel/centOS/7.2/ to
components/kernel/centOS/7.2/SOUCES. This will make easier to copy them to the
biuld and not confuse them w/ spec and other special files that we may have in
the future.
    - Yaml files were slightly refactored to include data regarding the place
where the files necessary to build the RPMs were stored. Package class obviously
was target of a similar refactoring.
    - distro and centOS classes were refactored, the distro class will start the
build process, there is no need for a specialized method be used since it should
not diverge from distro to distro.
    - Repository class now also create archives for the projects. Unfortunatelly
CentOS's pygit2 doesn't support archives w/ prefix parameter and thus the git
archive was made calling it though subprocess.Popen.
    - Added ./build directory to gitignore. That directory will be used to
receive intermidiate tar files srpm and final rpms besides the build logs and
such.