Fix kernel spec
So CentOS doesn't have base-rpm-config it has system-rpm-config, so let's
replace it.
Second there is a. imho, bug in the spec that creates the signmodules macro but
inside a conditional never reached which means that the macro was never created.
Instead of creating outside that scope, I prefered to adapt the if condition to
check for its existence using the syntax:
+%if 0%{?signmodules:0}