Setup result directory of the build and related options
Two new options were included: --keep-builddir in order to help in an eventual
debug of the build process (defaults to false) and --result-dir wihch allow user
to select destination fo the built packages (defaults to ./result directory
which was also included on gitignore for it).

Also refactored utils software discovery functionality to config.py in order to
avoid circular imports and removed reference to dependencies between built
packages, that will be on a next patch.