Add support for dependencies using scheduler
Added a very primitive schedules that one valid order to build the packages with
dependencies that are not solved or found by the distro itself.

    - Libseccomp's
    Moved libseccomp's files into a dependencies directory and added a line to
the spec to execute the autogen.sh which create files necessary to the build.
    - QEMU's
    Added libseccomp dependency.

Other changes included refactors and dependency/build dependency package
setup.

Fixes #3 and Fixes #15