Do not use root user to run the scripts
Instead of using root this patch proposes to use a user in mock's group and for
such it also includes a setup_environment.py tool that creates a user in that
groups OR add an existent user to it, this one though expects you to use root,
for instance:

 # sudo python setup_environment.py maurosr
This creates the user maurosr, if it doesn't exist, and add it to mock group.

It also setup the **default** directories for log and repositories for that
user.

Fixes #21