Fix scripts failing to produce results with newly created user (#29)
* Do not limit package choices in argument parser

The available packages are only discovered after running the complete build
for the first time. This prevented any packages from being built alone in the
first run, which is not desired.

* Make the path to mock binary explicit

If the binary were searched in the user PATH, the binary in /usr/sbin could be
used instead, requiring the user to have root privilege.