Add upgrade_versions.py tool
This tool is used to get newer commits withing the branch described in every
project's yaml.

This is very helpful for CI environment where once per week we checkout the HEAD
of our projects's branch to produce an updated weekly build.

Note: currently this tools only work for projects we get from git and the ones
we actively develop in order to offer the best virtualization technology for
Open Power Systems:
 - qemu
 - kernel
 - libvirt
 - kimchi
 - ginger
 - gingerbase
 - wok
 - sos
 - SLOF