Raise exceptions instead of returning error code in BuildManager
Internal classes should not have to deal with error codes, which are
responsability of the outer script. Exceptions are also easier to catch
and re-raise (this will be done in a next patch, which splits the
manager module).

The build method had no return code, so it made no sense to return it.