Remove categorization of dependency packages
All packages must be delivered, so we don't need to make a distinction
of which are dependencies or build dependencies when building them. It
just needs to be mentioned on the dependent package.
This also simplifies the directory structure of the versions repository
and solves the issue we could have when a package is both a dependency
and a build dependency.
The old directory structure, however, should still be supported for
earlier versions compatibility.