Update mock's tmpfs plugin parameters with realistic values
We had a misconception on the meaning of the parameters to the tmpfs
plugin. The "required_ram_mb" parameter must be greater than the
"max_fs_size" parameter, to make sure the system has enough RAM to mount
the whole tmpfs. Otherwise, the build may try to use more memory than
what is available and crash with an "out of memory" failure.

We've found out that guests running the whole build grow their
expandable virtual disks to aproximately 30GB, so the max file system
size must be at least that. With a little slack and adding 8GiB for
"normal" RAM usage, we get to the defined value of approximately 40GiB
of required RAM to enable the tmpfs plugin.