Make repository cloning/checkout lazy
This is to prepare for concurrent build script executions.

Since the source code repository is a shared resource among builds on
the same system, if the repositories were all cloned and checked out
before any package was built, we wouldn't gain much by executing build
scripts concurrently because the second process would have to wait for
the first to finish using the repository (archiving its contents) to be
able to modify the repositories directories (e.g. checking out a
different commit).
By cloning/checking out only when the package will start being built, we
reduce the time that the shared repository is in use by a given process.