Use lock files to allow concurrent build script executions
Using lock files, we are able to mark when the packages' shared resources
are being manipulated (the repository), so a second process can safely
execute operations on other packages and stall execution when it reaches
a locked package.
Since the whole package build is inside the critical section, this
creates a pipeline where only one process deals with a package at a
time.