Fix upgrade-versions command
When migrating from the old yaml sources format to the new one, the
commit that modified the upgrade_versions module
(1161796bbb0408eb981fbebf25c79bdd107b3f65) was incomplete, preventing
the spec release bumps.

The spec release bump and git commit logging were broken since commit
0839c75a5c7f528e7ca090d0542b04e843f20902. Only version updates were
working, but because the release bump was not even happening, the issue
was not observed.