Reorganize mock configuration files directory structure
Old structure contained an 'extras' directory which is only used for Mock
configuration files. As we do not have any 3rd party software config files
besides the Mock ones, this structure is too generic for our current needs,
making it difficult for a beginner to locate the Mock configuration files
in the git repository. The new structure makes that easier.