Anticipate the package's lock file unlock
After the repository is check out at the specified commit and archived
to a .tar.gz file, it is no longer used, so the package's lock can be
safely released.
To make it easier to spot the related lock and unlock calls and avoid
breaking the build function in too many public functions, I've
externalized the _prepare() function and renamed it to
prepare_sources(), allowing the lock to be released after the archive is
ready and the extra build files are copied to the build directory, which
should take a very small time compared to the actual build.
This allows the same package to be built by multiple processes in
parallel, while avoiding the need to clone the repository multiple
times.