Fix parsing of arguments for the build-iso command
It was getting the mock_config value from the "default" node in the
config.yaml file, because it has the same name as the attribute in the
"iso" node.

By overriding the "default" configurations with the ones from specific
subcommands, we make sure to get the right options.

Regression introduced by commit 45c167698fccea2cfede419493aaa7905feae333