Add commit_updates and push_updates parameters to upgrade-versions option
This is useful for developing and for giving the user the opportunity
of checking the updates before committing to the local git repository and
pushing to the remote git repository.

Also, before this patch, the name and email of the user written to the new
change log entry and set as commit author/email were called "committer name"
and "commiter email". As comitting is now optional, they were renamed to
"updater name" and "updater email".