Update local repository URL when it changes
When a local repository was found, the GitRepository object was created
from it and the remote repository URL passed to the get_git_repository
function was ignored. This lead to problems when the URL changed from
one execution to another, because a particular commit of interest could
be unavailable in the previously cloned repository.

This was particularly annoying for the versions repository, which can be
easily changed from the command line and changed frequently during our
tests. It would be even harder to debug for a user that wanted to change
a particular package's URL in the yaml file.