Copy files instead of moving in Mock._save_rpm method
We have an option --keep-builddir that keeps the build directory after
a successful package build. It might be confusing for the user if we
remove some build artifacts from this directory when the build finishes.

Also, this will allow us to reuse this method when we create a build
results cache.