Move selection of package builder and its conf file to BuildManager
Before, the selection of the package builder and its configuration file was
done by the Distribution class. This patch moves this reponsibility to the
facade BuildManager class and restricts the Distribution class responsibilities
to keeping the Linux distribution name, version, arch, endianness and checking
if it is supported by the build system. The purpose is to make the class design
simpler to understand.

Also, move PackageBuilder class to a separate Python module.