Remove build_results_dir attribute from Package class
This attribute is inferred from an option that is only available in the
build-package command and so it breaks the upgrade-versions command.

Since the only user of this attribute is the Mock class, that copies
the actual build results to this directory, it makes sense to create
this path there.

Regression introduced in commit f20d7166870e918c3976fb6846feca688ea4df44