Add Host OS package environments to ISO installation
Those environments contain groups of packages and appear to the user
in the "Software selection" screen, along with the default CentOS ones,
which are merged from another comps.xml available remotely. They must
contain the base packages groups that are installed in a "Minimal
install", along with a group containing our built packages.