Remove comment about moving config/host_os.yaml to /etc
It was recently decided by the team that all build system files
should be inside the builds git repository. The motivation of
this is to make the build system more auto-contained and
be easily usable by several users in the same system.

Given this decision, it does not make sense to move the file to
/etc anymore.