Make package_data a temporary variable of Package
This variable is initialized to the dict read from the YAML file, and was
previously held as a class attribute. It's values, however, are parsed to
class attributes, and they could become outdated if some of those
attributes were modified. To avoid issues in the future, let's remove
this unnecessary attribute.