Add powerpc-utils to ISO repo
Anaconda needs(1) the 'nvram' command from the powerpc-utils package
which used to be pulled as a dependency of ppc64-diag.

With "78ea1a2 Remove ppc64-diag", powerpc-utils is no longer present
in the ISO rootfs.

This patch adds powerpc-utils and ppc64-utils explicitly to the
repository provided to lorax (the latter is the metapackage installed
by lorax in the rootfs).

1 - 0c46bf244 Update ppc SMS bios after installation